requirements.txt
alabaster==0.7.12
argh==0.26.2
antlr4-python2-runtime==4.7.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
bleach==3.1.0
bumpversion==0.5.3
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
configparser==3.7.4
contextlib2==0.5.5
coverage==4.5.1
docutils==0.14
enum34==1.1.6
filelock==3.0.12
flake8==3.5.0
funcsigs==1.0.2
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==5.0.0
packaging==19.0
pathlib2==2.3.4
pathtools==0.1.2
pkginfo==1.5.0.1
pluggy==0.12.0
py==1.8.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.4.2
pyparsing==2.4.0
pytest==3.8.2
pytest-runner==4.2
pytz==2019.1
PyYAML==5.1.1
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
scandir==1.10.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.2
toml==0.10.0
tox==3.5.2
tqdm==4.32.2
twine==1.13.0
typing==3.7.4
urllib3==1.25.3
virtualenv==16.6.1
watchdog==0.9.0
webencodings==0.5.1
zipp==0.5.2
